+++ components/decoder/fdkaac/fdkaac.cpp
@@ -1,5 +1,5 @@
  /* BoCA - BonkEnc Component Architecture
-  * Copyright (C) 2007-2014 Robert Kausch <robert.kausch@bonkenc.org>
+  * Copyright (C) 2007-2015 Robert Kausch <robert.kausch@bonkenc.org>
   *
   * This program is free software; you can redistribute it and/or
   * modify it under the terms of the "GNU General Public License".
@@ -202,7 +202,7 @@ Error BoCA::DecoderFDKAAC::GetStreamInfo(const String &streamURI, Track &track)
 				format.rate	= streamInfo->sampleRate;
 				format.channels = streamInfo->numChannels;
 
-				track.length	= Math::Round(ex_MP4GetTrackDuration(mp4File, mp4Track) * format.rate / ex_MP4GetTrackTimeScale(mp4File, mp4Track));
+				track.length	= Math::Round(ex_MP4GetTrackDuration(mp4File, mp4Track) * Float(format.rate / ex_MP4GetTrackTimeScale(mp4File, mp4Track)));
 
 				format.bits	= 16;
 			}
@@ -222,7 +222,7 @@ Error BoCA::DecoderFDKAAC::GetStreamInfo(const String &streamURI, Track &track)
 					String			 valueString = (char *) items->elements[0].dataList.elements[0].value;
 					const Array<String>	&values	     = valueString.Trim().Explode(" ");
 
-					track.length = Math::Round((Int64) Number::FromHexString(values.GetNth(3)) * format.rate / ex_MP4GetTrackTimeScale(mp4File, mp4Track));
+					track.length = Math::Round((Int64) Number::FromHexString(values.GetNth(3)) * Float(format.rate / ex_MP4GetTrackTimeScale(mp4File, mp4Track)));
 
 					String::ExplodeFinish();
 				}
@@ -439,7 +439,7 @@ Bool BoCA::DecoderFDKAAC::Activate()
 				String			 valueString = (char *) items->elements[0].dataList.elements[0].value;
 				const Array<String>	&values	     = valueString.Trim().Explode(" ");
 
-				delaySamples	 = Math::Round((Int64) Number::FromHexString(values.GetNth(1)) * track.GetFormat().rate / ex_MP4GetTrackTimeScale(mp4File, mp4Track));
+				delaySamples	 = Math::Round((Int64) Number::FromHexString(values.GetNth(1)) * Float(track.GetFormat().rate / ex_MP4GetTrackTimeScale(mp4File, mp4Track)));
 				delaySamplesLeft = delaySamples;
 
 				String::ExplodeFinish();
